refact: reorganize test package
