+++ src/test/kotlin/smalltalk/backend/apply/repository/RoomRepositoryTest.kt
@@ -1,6 +1,5 @@
-package smalltalk.backend.apply.infrastructure.repository.room
+package smalltalk.backend.apply.repository
 
-import io.github.oshai.kotlinlogging.KotlinLogging
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.ExpectSpec
 import io.kotest.matchers.collections.*
@@ -13,7 +12,7 @@ import smalltalk.backend.config.redis.RedisConfig
 import smalltalk.backend.exception.room.situation.FullRoomException
 import smalltalk.backend.exception.room.situation.RoomNotFoundException
 import smalltalk.backend.infrastructure.repository.room.*
-import smalltalk.backend.support.EnableTestContainers
+import smalltalk.backend.support.EnableTestContainer
 import smalltalk.backend.support.spec.afterRootTest
 import smalltalk.backend.util.jackson.ObjectMapperClient
 
@@ -21,9 +20,10 @@ import smalltalk.backend.util.jackson.ObjectMapperClient
     classes = [RedisConfig::class, RoomRedisFunctionsLoader::class, RedissonRoomRepository::class, ObjectMapperClient::class]
 )
 @EnableConfigurationProperties(value = [RoomYamlProperties::class])
-@EnableTestContainers
-class RoomRepositoryTest(private val roomRepository: RoomRepository) : ExpectSpec({
-    val logger = KotlinLogging.logger { }
+@EnableTestContainer
+class RoomRepositoryTest(
+    private val roomRepository: RoomRepository,
+) : ExpectSpec({
 
     expect("채팅방을 저장한다") {
         roomRepository.save(NAME).run {
